+++ noodles-refget/examples/refget_download_sequence.rs
@@ -5,10 +5,9 @@
 use std::{
     env,
     io::{self, Write},
-    ops::Bound,
 };
 
-use noodles_core::Position;
+use noodles_core::region::Interval;
 use noodles_refget as refget;
 
 #[tokio::main]
@@ -17,21 +16,13 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     let base_url = args.next().expect("missing base URL").parse()?;
     let id = args.next().expect("missing ID");
+    let interval: Option<Interval> = args.next().map(|s| s.parse()).transpose()?;
 
     let client = refget::Client::new(base_url);
 
     let mut request = client.sequence(id);
 
-    if let Some(raw_start) = args.next() {
-        let start_bound: Bound<Position> = raw_start.parse().map(Bound::Included)?;
-
-        let end_bound = args
-            .next()
-            .map(|s| s.parse().map(Bound::Included))
-            .transpose()?
-            .unwrap_or(Bound::Unbounded);
-
-        let interval = (start_bound, end_bound);
+    if let Some(interval) = interval {
         request = request.set_interval(interval);
     }